Feature: Implementación de filtro paginado Se agrega filtro base paginado Se agregan dominios, entidades y request necesarios Nuevo metodo RouterRest para filtro Se agrega caso de uso para consultar y enriquecer data

+++ domain/model/src/main/java/co/com/bancolombia/model/loantype/LoanType.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 @Builder(toBuilder = true)
 public class LoanType {
 
-    private Integer idLoanType;
+    private Long idLoanType;
     private String name;
     private BigDecimal minAmount;
     private BigDecimal maxAmount;
