+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/bancolombia/r2dbc/entity/LoanRequestEntity.java
@@ -32,8 +32,8 @@ public class LoanRequestEntity {
     private String email;
 
     @Column("estado_id")
-    private Integer statusId;
+    private Long statusId;
 
     @Column("tipo_prestamo_id")
-    private Integer loanTypeId;
+    private Long loanTypeId;
 }
